chore: adds shrine contracts
