+++ packages/on-chain/contracts/astrodrop/lib/Ownable.sol
@@ -1,0 +1,54 @@
+// SPDX-License-Identifier: AGPL-3.0
+
+pragma solidity ^0.8.11;
+
+abstract contract Ownable {
+    error Ownable_NotOwner();
+    error Ownable_NewOwnerZeroAddress();
+
+    address private _owner;
+
+    event OwnershipTransferred(
+        address indexed previousOwner,
+        address indexed newOwner
+    );
+
+    /// @dev Initializes the contract setting the deployer as the initial owner.
+    function __Ownable_init(address initialOwner) internal {
+        _transferOwnership(initialOwner);
+    }
+
+    /// @dev Returns the address of the current owner.
+    function owner() public view virtual returns (address) {
+        return _owner;
+    }
+
+    /// @dev Throws if called by any account other than the owner.
+    modifier onlyOwner() {
+        if (owner() != msg.sender) revert Ownable_NotOwner();
+        _;
+    }
+
+    /// @dev Leaves the contract without owner. It will not be possible to call
+    /// `onlyOwner` functions anymore. Can only be called by the current owner.
+    /// NOTE: Renouncing ownership will leave the contract without an owner,
+    /// thereby removing any functionality that is only available to the owner.
+    function renounceOwnership() public virtual onlyOwner {
+        _transferOwnership(address(0));
+    }
+
+    /// @dev Transfers ownership of the contract to a new account (`newOwner`).
+    /// Can only be called by the current owner.
+    function transferOwnership(address newOwner) public virtual onlyOwner {
+        if (newOwner == address(0)) revert Ownable_NewOwnerZeroAddress();
+        _transferOwnership(newOwner);
+    }
+
+    /// @dev Transfers ownership of the contract to a new account (`newOwner`).
+    /// Internal function without access restriction.
+    function _transferOwnership(address newOwner) internal virtual {
+        address oldOwner = _owner;
+        _owner = newOwner;
+        emit OwnershipTransferred(oldOwner, newOwner);
+    }
+}
