+++ packages/on-chain/package.json
@@ -17,6 +17,7 @@
     "@nomicfoundation/hardhat-toolbox": "^2.0.0",
     "@nomiclabs/hardhat-ethers": "^2.2.1",
     "@nomiclabs/hardhat-etherscan": "^3.1.3",
+    "@openzeppelin/contracts": "^4.8.0",
     "@typechain/ethers-v5": "^10.2.0",
     "@typechain/hardhat": "^6.1.5",
     "@types/jest": "^29.2.4",
@@ -26,6 +27,7 @@
     "hardhat-deploy": "^0.11.22",
     "hardhat-gas-reporter": "^1.0.9",
     "solidity-coverage": "^0.8.2",
+    "solmate": "^6.6.1",
     "typechain": "^8.1.1"
   }
 }
